# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tumbas POS API",
    "description": "Minimal point of sale API for products, checkout, sales history, and sales summary.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
...
  "paths": {
...
    "/api/v1/inventory/opname/{id}": {
      "get": {
        "tags": [
          "Opname"
        ],
        "summary": "Show opname details",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opname details."
          }
        }
      },
      "put": {
        "tags": [
          "Opname"
        ],
        "summary": "Finalize draft opname",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opname finalized."
          }
        }
      },
      "delete": {
        "tags": [
          "Opname"
        ],
        "summary": "Delete draft opname",
        "description": "Deletes a draft stock opname sheet. Completed opname cannot be deleted. Restricted to manage_inventory permission.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Draft opname deleted."
          }
        }
      }
    },
...
    "/api/v1/products/{product}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get product detail",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProductId"
          }
        ],
        "responses": {
          "200": {
            "description": "Product retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Replace or update product",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProductId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              },
              "example": {
                "nama": "Teh Botol",
                "merek": "Sosro",
                "stok": 20,
                "harga": 5500
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Products"
        ],
        "summary": "Partially update product",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProductId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              },
              "example": {
                "stok": 20
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete product",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProductId"
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the Sanctum plainTextToken retrieved from login."
      }
    },
    "parameters": {
      "ProductId": {
        "name": "product",
        "in": "path",
        "description": "Product ID.",
        "required": true,
        "schema": {
          "type": "integer",
          "example": 1
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Validation error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Product": {
        "type": "object",
        "required": [
          "id",
          "nama",
          "merek",
          "stok",
          "harga",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "nama": {
            "type": "string",
            "example": "Teh Botol"
          },
          "merek": {
            "type": "string",
            "nullable": true,
            "example": "Sosro"
          },
          "stok": {
            "type": "integer",
            "minimum": 0,
            "example": 12
          },
          "harga": {
            "type": "integer",
            "minimum": 0,
            "example": 5000
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "ProductUpdate": {
        "type": "object",
        "properties": {
          "nama": {
            "type": "string",
            "maxLength": 255,
            "example": "Teh Botol"
          },
          "merek": {
            "type": "string",
            "maxLength": 255,
            "nullable": true,
            "example": "Sosro"
          },
          "stok": {
            "type": "integer",
            "minimum": 0,
            "example": 20
          },
          "harga": {
            "type": "integer",
            "minimum": 0,
            "example": 5500
          }
        }
      },
...
      "ValidationError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "The nama field is required."
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "nama": [
                "The nama field is required."
              ]
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Not Found"
          }
        }
      },
...
    }
  }
}